+++ Changelog.md
@@ -5,6 +5,7 @@
 - Fix Audio decoding method *Pipe* causes crash (Dendraspis, #340)
 - Fix parameters when audio converting to AAC with FFmpeg (Dendraspis, Patman86, #341)
 - Improved Opus encoder (Patman86)
+- NeroAAC removal
 
 
 2.1.4.6 Beta (2020-09-13)